Add a mdev rule to rename hw_random to hwrng
When mdev is invoked to populate /dev (mdev -s), it creates the node
for the hardware random number generator (major 10, minor 183) with
the incorrect name. This is because the linux kernel uses 'hw_random'
for the /sys entry and 'hwrng' for the /dev entry. As a result, when
mdev scans /sys/class, it finds the entry
'/sys/class/misc/hw_random/dev' and creates /dev/hw_random.

When combined with the kernel option 'CONFIG_DEVTMPFS', this results
in having two nodes under /dev (namely 'hw_random' and 'hwrng) that
point to the same device. This currently breaks sys-apps/rng-tools
(#481254).

The use of 'hw_random' has been obselete for 6 years [1]. Kernel
exposes no mechanism for mdev to figure out the node name
automatically. Utilize 'mdev.conf' to manually rename 'hw_random' to
'hwrng'.

[1] http://marc.info/?l=linux-crypto-vger&m=144249767024990&w=2

Signed-off-by: Gokturk Yuksek <gokturk@binghamton.edu>