Disable unnecessary udev timeout in zpool import
This is a regression that was introduced in sys-fs/zfs-kmod-0.7.0.
It was originally fixed by 2eb1d04cfbfa397b58a0b388f8ed28688fd114d8, but
this neglected to handle the case where booting is done via arguments
such as "root=ZFS" or "root=ZFS=rpool/ROOT/gentoo" on the kernel
commandline. This handles it.

Signed-off-by: Richard Yao <ryao@gentoo.org>