zfs: allow unlock multiple encrypted pools
Added 2 new values for `dozfs`.

`loadkeyall` will use `zfs load-key -a` to ask all passphrases, not just
root.

`unifiedkey` will try to use a single key to unlock all pools.
It must be used with `loadkeyall`. It does not need to be able to unlock
all pools, but it must be able to unlock the root pool.

Also `unlock-zfs` support `-a` for all pools, and `-u` for unified key.

Signed-off-by: Rin Cat (鈴猫) <rincat@rincat.dev>